Move opaque check function to Graphic impl
